- split the relayedTx tests in multiple sub-packages
