+++ integrationTests/multiShard/relayedTx/attestationContract/attestation.wasm
